+++ history.md
@@ -3,6 +3,10 @@
 
 -     search ieeexplore (Wed Jul 8 16:20:42 2015 +0200) - [view](../../commit/4b16b73d41a3a573ab0cd02aa56dd808d9b9d38f) 
 
+# Refactorings
+
+-     move ieee search into its own module (Wed Jul 8 17:43:12 2015 +0200) - [view](../../commit/e58659bfdc0985a4db87cd843fa1381825590b01) 
+
 # Changes to the build process
 
 -     move sources into the appropriate source directory (Wed Jul 8 16:45:44 2015 +0200) - [view](../../commit/3beb254e8013ba0b580947d916c44e267e187fc3)